#65 - Refactor user service to use separate router for CRUD operations and improve MongoDB connection logging
